Add type constraint to please the 4.04.0 compiler
Because of an obscure fix to an obscure bug in 4.03.0 and earlier, the 4.04.0 compiler will reject this code if you don't explicitly state the type of the elements (not) stored into the `list` variable. This fix should be compatible with all earlier OCaml versions.

See http://caml.inria.fr/mantis/view.php?id=6752